+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/PendingPromise.kt
@@ -1,7 +1,7 @@
 package expo.modules.imagepicker
 
 import android.os.Bundle
-import expo.modules.core.Promise
+import expo.modules.kotlin.Promise
 
 /**
  * Class that represents a promise, which will not resolve immediately but saves results to the [PickerResultsStore].
@@ -24,7 +24,7 @@ class PendingPromise(
     throw IllegalArgumentException("Can not resolve 'DestroyedPromise' with anything else then 'Bundle'.")
   }
 
-  override fun reject(code: String, message: String, e: Throwable?) {
+  override fun reject(code: String, message: String?, e: Throwable?) {
     pickerResultsStore.addPendingResult(
       Bundle().apply {
         putString("code", code)