+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/tasks/ImageResultTask.kt
@@ -4,16 +4,17 @@ import android.content.ContentResolver
 import android.net.Uri
 import android.os.Bundle
 import android.util.Base64
-import android.util.Log
 import androidx.exifinterface.media.ExifInterface
-import expo.modules.core.Promise
-import expo.modules.core.errors.ModuleDestroyedException
 import expo.modules.imagepicker.ExifDataHandler
-import expo.modules.imagepicker.ImagePickerConstants
-import expo.modules.imagepicker.ImagePickerConstants.exifTags
+import expo.modules.imagepicker.FailedToExtractVideoMetadataException
+import expo.modules.imagepicker.FailedToWriteFileException
+import expo.modules.imagepicker.ImagePickerConstants.EXIF_TAGS
+import expo.modules.imagepicker.ImagePickerMediaResponse
+import expo.modules.imagepicker.UnknownException
 import expo.modules.imagepicker.exporters.ImageExporter
 import expo.modules.imagepicker.exporters.ImageExporter.Listener
 import expo.modules.imagepicker.fileproviders.FileProvider
+import expo.modules.kotlin.Promise
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.suspendCancellableCoroutine
@@ -70,36 +71,25 @@ open class ImageResultTask(
         val exif = getExifData()
         val imageExporterHandler = object : Listener {
           override fun onResult(out: ByteArrayOutputStream?, width: Int, height: Int) {
-            val response = Bundle().apply {
-              putString("uri", Uri.fromFile(outputFile).toString())
-              putInt("width", width)
-              putInt("height", height)
-              putBoolean("cancelled", false)
-              putString("type", "image")
-
-              out?.let {
-                putString("base64", Base64.encodeToString(it.toByteArray(), Base64.NO_WRAP))
-              }
-              exif?.let {
-                putBundle("exif", it)
-              }
-            }
+            val response = ImagePickerMediaResponse.Image(
+              uri = Uri.fromFile(outputFile).toString(),
+              width = width,
+              height = height,
+              base64 = out?.let { Base64.encodeToString(it.toByteArray(), Base64.NO_WRAP) },
+              exif = exif
+            )
             promise.resolve(response)
           }
 
           override fun onFailure(cause: Throwable?) {
-            promise.reject(ImagePickerConstants.ERR_CAN_NOT_SAVE_RESULT, ImagePickerConstants.CAN_NOT_SAVE_RESULT_MESSAGE, cause)
+            promise.reject(FailedToWriteFileException(outputFile, cause))
           }
         }
         imageExporter.export(uri, outputFile, imageExporterHandler)
-      } catch (e: ModuleDestroyedException) {
-        Log.i(ImagePickerConstants.TAG, ImagePickerConstants.COROUTINE_CANCELED, e)
-        promise.reject(ImagePickerConstants.COROUTINE_CANCELED, e)
       } catch (e: IOException) {
-        promise.reject(ImagePickerConstants.ERR_CAN_NOT_EXTRACT_METADATA, ImagePickerConstants.CAN_NOT_EXTRACT_METADATA_MESSAGE, e)
+        promise.reject(FailedToExtractVideoMetadataException(e))
       } catch (e: Exception) {
-        Log.e(ImagePickerConstants.TAG, ImagePickerConstants.UNKNOWN_EXCEPTION, e)
-        promise.reject(ImagePickerConstants.UNKNOWN_EXCEPTION, e)
+        promise.reject(UnknownException(e))
       }
     }
   }
@@ -108,16 +98,16 @@ open class ImageResultTask(
   private fun readExif() = Bundle().apply {
     contentResolver.openInputStream(uri)?.use { input ->
       val exifInterface = ExifInterface(input)
-      exifTags.forEach { (type, name) ->
-        if (exifInterface.getAttribute(name) != null) {
-          when (type) {
-            "string" -> putString(name, exifInterface.getAttribute(name))
-            "int" -> putInt(name, exifInterface.getAttributeInt(name, 0))
-            "double" -> putDouble(name, exifInterface.getAttributeDouble(name, 0.0))
-          }
+      EXIF_TAGS.forEach { (type, tag) ->
+        if (exifInterface.getAttribute(tag) == null) {
+          return@forEach
+        }
+        when (type) {
+          "string" -> putString(tag, exifInterface.getAttribute(tag))
+          "int" -> putInt(tag, exifInterface.getAttributeInt(tag, 0))
+          "double" -> putDouble(tag, exifInterface.getAttributeDouble(tag, 0.0))
         }
       }
-
       // Explicitly get latitude, longitude, altitude with their specific accessor functions.
       exifInterface.latLong?.let { latLong ->
         putDouble(ExifInterface.TAG_GPS_LATITUDE, latLong[0])